1、将自动生成的Short改为integer 2、解决Oracle连接包的问题

+++ src/main/java/com/qianfeng/wkrent/dao/BrandMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Brand;
 
 public interface BrandMapper {
-    int deleteByPrimaryKey(Short brandId);
+    int deleteByPrimaryKey(Integer brandId);
 
     int insert(Brand record);
 
     int insertSelective(Brand record);
 
-    Brand selectByPrimaryKey(Short brandId);
+    Brand selectByPrimaryKey(Integer brandId);
 
     int updateByPrimaryKeySelective(Brand record);
