+++ src/main/java/com/qianfeng/wkrent/dao/CarMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Car;
 
 public interface CarMapper {
-    int deleteByPrimaryKey(Short carId);
+    int deleteByPrimaryKey(Integer carId);
 
     int insert(Car record);
 
     int insertSelective(Car record);
 
-    Car selectByPrimaryKey(Short carId);
+    Car selectByPrimaryKey(Integer carId);
 
     int updateByPrimaryKeySelective(Car record);
