+++ src/main/java/com/qianfeng/wkrent/dao/ComboMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Combo;
 
 public interface ComboMapper {
-    int deleteByPrimaryKey(Short comboId);
+    int deleteByPrimaryKey(Integer comboId);
 
     int insert(Combo record);
 
     int insertSelective(Combo record);
 
-    Combo selectByPrimaryKey(Short comboId);
+    Combo selectByPrimaryKey(Integer comboId);
 
     int updateByPrimaryKeySelective(Combo record);
