+++ src/main/java/com/qianfeng/wkrent/dao/ComplaintTypeMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.ComplaintType;
 
 public interface ComplaintTypeMapper {
-    int deleteByPrimaryKey(Short typeId);
+    int deleteByPrimaryKey(Integer typeId);
 
     int insert(ComplaintType record);
 
     int insertSelective(ComplaintType record);
 
-    ComplaintType selectByPrimaryKey(Short typeId);
+    ComplaintType selectByPrimaryKey(Integer typeId);
 
     int updateByPrimaryKeySelective(ComplaintType record);
