+++ src/main/java/com/qianfeng/wkrent/dao/DayPriceMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.DayPrice;
 
 public interface DayPriceMapper {
-    int deleteByPrimaryKey(Short dpId);
+    int deleteByPrimaryKey(Integer dpId);
 
     int insert(DayPrice record);
 
     int insertSelective(DayPrice record);
 
-    DayPrice selectByPrimaryKey(Short dpId);
+    DayPrice selectByPrimaryKey(Integer dpId);
 
     int updateByPrimaryKeySelective(DayPrice record);
