+++ src/main/java/com/qianfeng/wkrent/dao/InstructionMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Instruction;
 
 public interface InstructionMapper {
-    int deleteByPrimaryKey(Short instructionId);
+    int deleteByPrimaryKey(Integer instructionId);
 
     int insert(Instruction record);
 
     int insertSelective(Instruction record);
 
-    Instruction selectByPrimaryKey(Short instructionId);
+    Instruction selectByPrimaryKey(Integer instructionId);
 
     int updateByPrimaryKeySelective(Instruction record);
