+++ src/main/java/com/qianfeng/wkrent/dao/MonthPriceMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.MonthPrice;
 
 public interface MonthPriceMapper {
-    int deleteByPrimaryKey(Short mpId);
+    int deleteByPrimaryKey(Integer mpId);
 
     int insert(MonthPrice record);
 
     int insertSelective(MonthPrice record);
 
-    MonthPrice selectByPrimaryKey(Short mpId);
+    MonthPrice selectByPrimaryKey(Integer mpId);
 
     int updateByPrimaryKeySelective(MonthPrice record);
