+++ src/main/java/com/qianfeng/wkrent/dao/PermissionMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Permission;
 
 public interface PermissionMapper {
-    int deleteByPrimaryKey(Short permissionId);
+    int deleteByPrimaryKey(Integer permissionId);
 
     int insert(Permission record);
 
     int insertSelective(Permission record);
 
-    Permission selectByPrimaryKey(Short permissionId);
+    Permission selectByPrimaryKey(Integer permissionId);
 
     int updateByPrimaryKeySelective(Permission record);
