+++ src/main/java/com/qianfeng/wkrent/dao/PlaceMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Place;
 
 public interface PlaceMapper {
-    int deleteByPrimaryKey(Short placeId);
+    int deleteByPrimaryKey(Integer placeId);
 
     int insert(Place record);
 
     int insertSelective(Place record);
 
-    Place selectByPrimaryKey(Short placeId);
+    Place selectByPrimaryKey(Integer placeId);
 
     int updateByPrimaryKeySelective(Place record);
