+++ src/main/java/com/qianfeng/wkrent/dao/RentOrderMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.RentOrder;
 
 public interface RentOrderMapper {
-    int deleteByPrimaryKey(Short orderId);
+    int deleteByPrimaryKey(Integer orderId);
 
     int insert(RentOrder record);
 
     int insertSelective(RentOrder record);
 
-    RentOrder selectByPrimaryKey(Short orderId);
+    RentOrder selectByPrimaryKey(Integer orderId);
 
     int updateByPrimaryKeySelective(RentOrder record);
