+++ src/main/java/com/qianfeng/wkrent/dao/RoleMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Role;
 
 public interface RoleMapper {
-    int deleteByPrimaryKey(Short roleId);
+    int deleteByPrimaryKey(Integer roleId);
 
     int insert(Role record);
 
     int insertSelective(Role record);
 
-    Role selectByPrimaryKey(Short roleId);
+    Role selectByPrimaryKey(Integer roleId);
 
     int updateByPrimaryKeySelective(Role record);
