+++ src/main/java/com/qianfeng/wkrent/dao/RoomMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Room;
 
 public interface RoomMapper {
-    int deleteByPrimaryKey(Short roomId);
+    int deleteByPrimaryKey(Integer roomId);
 
     int insert(Room record);
 
     int insertSelective(Room record);
 
-    Room selectByPrimaryKey(Short roomId);
+    Room selectByPrimaryKey(Integer roomId);
 
     int updateByPrimaryKeySelective(Room record);
