+++ src/main/java/com/qianfeng/wkrent/dao/ScenicMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Scenic;
 
 public interface ScenicMapper {
-    int deleteByPrimaryKey(Short scenicId);
+    int deleteByPrimaryKey(Integer scenicId);
 
     int insert(Scenic record);
 
     int insertSelective(Scenic record);
 
-    Scenic selectByPrimaryKey(Short scenicId);
+    Scenic selectByPrimaryKey(Integer scenicId);
 
     int updateByPrimaryKeySelective(Scenic record);
