+++ src/main/java/com/qianfeng/wkrent/dao/TicketMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Ticket;
 
 public interface TicketMapper {
-    int deleteByPrimaryKey(Short ticketId);
+    int deleteByPrimaryKey(Integer ticketId);
 
     int insert(Ticket record);
 
     int insertSelective(Ticket record);
 
-    Ticket selectByPrimaryKey(Short ticketId);
+    Ticket selectByPrimaryKey(Integer ticketId);
 
     int updateByPrimaryKeySelective(Ticket record);
