+++ src/main/java/com/qianfeng/wkrent/dao/TripMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.Trip;
 
 public interface TripMapper {
-    int deleteByPrimaryKey(Short tripId);
+    int deleteByPrimaryKey(Integer tripId);
 
     int insert(Trip record);
 
     int insertSelective(Trip record);
 
-    Trip selectByPrimaryKey(Short tripId);
+    Trip selectByPrimaryKey(Integer tripId);
 
     int updateByPrimaryKeySelective(Trip record);
