+++ src/main/java/com/qianfeng/wkrent/dao/UserMapper.java
@@ -3,13 +3,13 @@ package com.qianfeng.wkrent.dao;
 import com.qianfeng.wkrent.dto.User;
 
 public interface UserMapper {
-    int deleteByPrimaryKey(Short userId);
+    int deleteByPrimaryKey(Integer userId);
 
     int insert(User record);
 
     int insertSelective(User record);
 
-    User selectByPrimaryKey(Short userId);
+    User selectByPrimaryKey(Integer userId);
 
     int updateByPrimaryKeySelective(User record);
